requirements.txt
pyobjc-framework-Quartz>=9.2
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyautogui>=0.9.54
easyocr>=1.7.1